refactor: DisplayName 의 문구를 BDD 에 걸맞게 수정
