+++ src/test/java/com/camping/legacy/acceptance/ReservationAcceptanceTest.java
@@ -1,7 +1,6 @@
 package com.camping.legacy.acceptance;
 
 import com.camping.legacy.support.TestBase;
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
@@ -9,18 +8,18 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
 import java.time.LocalDate;
-import java.util.Map;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CountDownLatch;
 import java.util.ArrayList;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("캠핑장 예약 시스템 통합 인수 테스트")
-class CampingReservationAcceptanceTest extends TestBase {
-    
-    @DisplayName("시나리오: 정상적인 예약 생성")
+@DisplayName("예약 시스템 인수 테스트")
+class ReservationAcceptanceTest extends TestBase {
+
+    @DisplayName("올바른 고객 정보와 날짜로 예약을 요청하면, 예약이 성공적으로 생성되고 6자리 확인코드가 발급된다")
     @Test
     void scenario_SuccessfulReservationCreation() {
         /*
@@ -35,29 +34,29 @@ class CampingReservationAcceptanceTest extends TestBase {
 
         // Background: 캠핑 사이트 "A001"(대형)과 "B001"(소형)이 등록되어 있다
         // 오늘 날짜는 현재 날짜이다
-        
+
         // Given: 고객 "김철수"의 전화번호는 "010-1234-5678"이다
         String customerName = "김철수";
         String phoneNumber = "010-1234-5678";
         LocalDate today = LocalDate.now();
         LocalDate startDate = today.plusDays(5);
         LocalDate endDate = startDate.plusDays(2);
-        
+
         // When: startDate부터 endDate까지 사이트 "1"을 예약 요청한다
         ExtractableResponse<Response> response = createReservation(customerName, phoneNumber, "1", startDate, endDate);
-        
+
         // Then: 예약이 성공적으로 생성된다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        
+
         // And: 6자리 확인코드가 발급된다
         String confirmationCode = response.jsonPath().getString("confirmationCode");
         assertThat(confirmationCode).isNotNull().hasSize(6);
-        
+
         // And: 예약 상태는 "CONFIRMED"이다
         assertThat(response.jsonPath().getString("status")).isEqualTo("CONFIRMED");
     }
-    
-    @DisplayName("시나리오: 30일 초과 예약 시도")
+
+    @DisplayName("30일을 초과한 날짜로 예약을 시도하면, '30일 이내 예약만 가능합니다'는 오류 메시지를 받는다")
     @Test
     void scenario_ReservationExceedsDateLimit() {
         /*
@@ -74,18 +73,18 @@ class CampingReservationAcceptanceTest extends TestBase {
         String customerName = "김철수";
         String phoneNumber = "010-1234-5678";
         LocalDate farFutureDate = LocalDate.now().plusDays(35);
-        
+
         // When: 30일 초과 날짜로 예약을 요청한다
         ExtractableResponse<Response> response = createReservation(customerName, phoneNumber, "1", farFutureDate, farFutureDate.plusDays(2));
-        
+
         // Then: 예약이 실패한다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        
+
         // And: "30일 이내 예약만 가능합니다" 오류 메시지를 받는다
         assertThat(response.jsonPath().getString("message")).contains("30일");
     }
-    
-    @DisplayName("시나리오: 과거 날짜 예약 시도")
+
+    @DisplayName("과거 날짜로 예약을 시도하면, '과거 날짜로 예약할 수 없습니다'는 오류 메시지를 받는다")
     @Test
     void scenario_PastDateReservationAttempt() {
         /*
@@ -102,18 +101,18 @@ class CampingReservationAcceptanceTest extends TestBase {
         String customerName = "김철수";
         String phoneNumber = "010-1234-5678";
         LocalDate pastDate = LocalDate.now().minusDays(1);
-        
+
         // When: 과거 날짜로 예약을 요청한다
         ExtractableResponse<Response> response = createReservation(customerName, phoneNumber, "1", pastDate, pastDate.plusDays(2));
-        
+
         // Then: 예약이 실패한다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        
+
         // And: "과거 날짜로 예약할 수 없습니다" 오류 메시지를 받는다
         assertThat(response.jsonPath().getString("message")).contains("과거");
     }
-    
-    @DisplayName("시나리오: 동시 예약 요청 시 하나만 성공")
+
+    @DisplayName("두 고객이 동시에 같은 사이트와 기간을 예약 요청하면, 하나의 예약만 성공하고 나머지는 충돌 오류로 실패한다")
     @Test
     void scenario_ConcurrentReservationOnlyOneSucceeds() throws Exception {
         /*
@@ -130,11 +129,11 @@ class CampingReservationAcceptanceTest extends TestBase {
         String siteId = "1";
         LocalDate startDate = LocalDate.now().plusDays(10);
         LocalDate endDate = startDate.plusDays(2);
-        
+
         CountDownLatch startLatch = new CountDownLatch(1);
         CountDownLatch doneLatch = new CountDownLatch(2);
         List<ExtractableResponse<Response>> responses = new ArrayList<>();
-        
+
         // When: 두 고객이 동시에 같은 사이트/기간을 예약 요청한다
         CompletableFuture<Void> future1 = CompletableFuture.runAsync(() -> {
             try {
@@ -148,7 +147,7 @@ class CampingReservationAcceptanceTest extends TestBase {
                 Thread.currentThread().interrupt();
             }
         });
-        
+
         CompletableFuture<Void> future2 = CompletableFuture.runAsync(() -> {
             try {
                 startLatch.await();
@@ -161,32 +160,32 @@ class CampingReservationAcceptanceTest extends TestBase {
                 Thread.currentThread().interrupt();
             }
         });
-        
+
         startLatch.countDown();
         doneLatch.await();
-        
+
         // Then: 하나의 예약만 성공한다
         long successCount = responses.stream()
                 .mapToInt(ExtractableResponse::statusCode)
                 .filter(status -> status == HttpStatus.CREATED.value())
                 .count();
         assertThat(successCount).isEqualTo(1L);
-        
+
         // And: 나머지 예약은 "이미 예약된 기간입니다" 오류로 실패한다
         long conflictCount = responses.stream()
                 .mapToInt(ExtractableResponse::statusCode)
                 .filter(status -> status == HttpStatus.CONFLICT.value())
                 .count();
         assertThat(conflictCount).isEqualTo(1L);
-        
+
         ExtractableResponse<Response> conflictResponse = responses.stream()
                 .filter(r -> r.statusCode() == HttpStatus.CONFLICT.value())
                 .findFirst()
                 .orElseThrow();
         assertThat(conflictResponse.jsonPath().getString("message")).contains("이미 예약");
     }
-    
-    @DisplayName("시나리오: 연박 예약 시 전체 기간 가용성 확인")
+
+    @DisplayName("이미 예약된 날짜가 중간에 포함된 기간으로 연박 예약을 시도하면, '해당 기간에 예약이 불가능합니다'는 오류 메시지를 받는다")
     @Test
     void scenario_MultiDayReservationFullPeriodAvailability() {
         /*
@@ -202,23 +201,23 @@ class CampingReservationAcceptanceTest extends TestBase {
         // Given: 사이트 "1"의 중간 날짜가 이미 예약되어 있다
         String siteId = "1";
         LocalDate conflictDate = LocalDate.now().plusDays(16);
-        
+
         // 중간 날짜에 기존 예약 생성
         createReservation("기존예약자", "010-0000-0000", siteId, conflictDate, conflictDate.plusDays(1));
-        
+
         // When: 기존 예약과 겹치는 기간으로 연박 예약을 요청한다
         LocalDate startDate = conflictDate.minusDays(1);
         LocalDate endDate = conflictDate.plusDays(2);
         ExtractableResponse<Response> response = createReservation("김철수", "010-1234-5678", siteId, startDate, endDate);
-        
+
         // Then: 예약이 실패한다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CONFLICT.value());
-        
+
         // And: "해당 기간에 예약이 불가능합니다" 오류 메시지를 받는다
         assertThat(response.jsonPath().getString("message")).contains("해당 기간");
     }
-    
-    @DisplayName("시나리오: 올바른 확인코드로 예약 취소")
+
+    @DisplayName("올바른 확인코드로 예약 취소를 요청하면, 예약 취소가 성공하고 상태가 CANCELLED로 변경된다")
     @Test
     void scenario_SuccessfulCancellationWithCorrectConfirmationCode() {
         /*
@@ -235,24 +234,24 @@ class CampingReservationAcceptanceTest extends TestBase {
         String customerName = "김철수";
         String phoneNumber = "010-1234-5678";
         LocalDate startDate = LocalDate.now().plusDays(20);
-        
+
         ExtractableResponse<Response> createResponse = createReservation(customerName, phoneNumber, "1", startDate, startDate.plusDays(2));
         Long reservationId = createResponse.jsonPath().getLong("id");
         String confirmationCode = createResponse.jsonPath().getString("confirmationCode");
-        
+
         // When: 올바른 확인코드로 예약 취소를 요청한다
         ExtractableResponse<Response> response = cancelReservation(reservationId, confirmationCode);
-        
+
         // Then: 예약 취소가 성공한다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         assertThat(response.jsonPath().getString("message")).contains("취소");
-        
+
         // And: 예약 상태가 "CANCELLED"로 변경된다
         ExtractableResponse<Response> getResponse = getReservation(reservationId);
         assertThat(getResponse.jsonPath().getString("status")).isEqualTo("CANCELLED");
     }
-    
-    @DisplayName("시나리오: 잘못된 확인코드로 예약 취소 시도")
+
+    @DisplayName("잘못된 확인코드로 예약 취소를 시도하면, '확인코드가 일치하지 않습니다'는 오류 메시지를 받는다")
     @Test
     void scenario_CancellationWithWrongConfirmationCode() {
         /*
@@ -266,21 +265,22 @@ class CampingReservationAcceptanceTest extends TestBase {
          */
 
         // Given: 고객 "김철수"가 확인코드로 예약을 완료했다
-        ExtractableResponse<Response> createResponse = createReservation("김철수", "010-1234-5678", "1", 
-                LocalDate.now().plusDays(22), LocalDate.now().plusDays(24));
+        ExtractableResponse<Response> createResponse = createReservation("김철수", "010-1234-5678", "1",
+                                                                         LocalDate.now().plusDays(22), LocalDate.now().plusDays(24)
+        );
         Long reservationId = createResponse.jsonPath().getLong("id");
-        
+
         // When: 잘못된 확인코드로 예약 취소를 요청한다
         ExtractableResponse<Response> response = cancelReservation(reservationId, "WRONG1");
-        
+
         // Then: 예약 취소가 실패한다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        
+
         // And: "확인코드가 일치하지 않습니다" 오류 메시지를 받는다
         assertThat(response.jsonPath().getString("message")).contains("확인코드");
     }
-    
-    @DisplayName("시나리오: 취소된 예약 사이트 재예약 가능")
+
+    @DisplayName("예약이 취소된 사이트에 새로운 예약을 요청하면, 예약이 성공적으로 생성된다")
     @Test
     void scenario_CancelledSiteCanBeReservedAgain() {
         /*
@@ -296,101 +296,22 @@ class CampingReservationAcceptanceTest extends TestBase {
         // Given: 사이트의 예약이 취소된 상태이다
         String siteId = "1";
         LocalDate reservationDate = LocalDate.now().plusDays(25);
-        
+
         ExtractableResponse<Response> createResponse = createReservation("김철수", "010-1234-5678", siteId, reservationDate, reservationDate.plusDays(1));
         Long reservationId = createResponse.jsonPath().getLong("id");
         String confirmationCode = createResponse.jsonPath().getString("confirmationCode");
-        
+
         // 예약 취소
         cancelReservation(reservationId, confirmationCode);
-        
+
         // When: 취소된 사이트에 새로 예약을 요청한다
         ExtractableResponse<Response> response = createReservation("이영희", "010-5555-5555", siteId, reservationDate, reservationDate.plusDays(1));
-        
+
         // Then: 예약이 성공적으로 생성된다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
     }
-    
-    @DisplayName("시나리오: 기간별 가용 사이트 검색")
-    @Test
-    void scenario_SearchAvailableSitesByPeriod() {
-        /*
-         * 누가(Who), 무엇을(What), 언제(When), 어디서(Where), 왜(Why), 어떻게(How)
-         * 누가(Who): 가용한 사이트를 찾려는 사용자
-         * 무엇을(What): 특정 기간에 가용한 사이트 목록을
-         * 언제(When): 사이트 예약이 있는 기간에
-         * 어디서(Where): 캠핑장 예약 시스템에서
-         * 왜(Why): 이미 예약된 사이트는 검색 결과에서 제외되어야 하기 때문에
-         * 어떻게(How): 이미 예약된 사이트가 있는 기간에 사이트 검색을 해서
-         */
 
-        // Given: 사이트 "1"이 특정 기간 예약되어 있다
-        String reservedSiteId = "1";
-        LocalDate startDate = LocalDate.now().plusDays(28);
-        LocalDate endDate = startDate.plusDays(3);
-        
-        createReservation("예약자", "010-0000-0000", reservedSiteId, startDate, endDate);
-        
-        // When: 해당 기간에 가용한 사이트를 검색한다
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .queryParam("startDate", startDate.toString())
-                .queryParam("endDate", endDate.toString())
-                .when().get("/api/sites/search")
-                .then().log().all()
-                .extract();
-        
-        // Then: 예약된 사이트는 검색 결과에 나타나지 않는다
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        List<Map<String, Object>> availableSites = response.jsonPath().getList("$");
-        
-        boolean containsReservedSite = availableSites.stream()
-                .anyMatch(site -> reservedSiteId.equals(String.valueOf(site.get("id"))));
-        assertThat(containsReservedSite).isFalse();
-    }
-    
-    @DisplayName("시나리오: 사이트 크기별 필터링")
-    @Test
-    void scenario_FilterSitesBySize() {
-        /*
-         * 누가(Who), 무엇을(What), 언제(When), 어디서(Where), 왜(Why), 어떻게(How)
-         * 누가(Who): 특정 크기의 사이트만 원하는 사용자
-         * 무엇을(What): 대형 사이트만을
-         * 언제(When): 특정 기간 동안
-         * 어디서(Where): 캠핑장 예약 시스템에서
-         * 왜(Why): 대형 사이트만 필요하기 때문에
-         * 어떻게(How): "대형" 크기 필터를 사용하여 사이트 검색을 해서
-         */
-
-        // Given: 대형 사이트와 소형 사이트가 있다
-        LocalDate searchStart = LocalDate.now().plusDays(29);
-        LocalDate searchEnd = searchStart.plusDays(2);
-        String sizeFilter = "대형";
-        
-        // When: "대형" 사이트만 검색한다
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .queryParam("startDate", searchStart.toString())
-                .queryParam("endDate", searchEnd.toString())
-                .queryParam("size", sizeFilter)
-                .when().get("/api/sites/search")
-                .then().log().all()
-                .extract();
-        
-        // Then: 대형 사이트만 검색 결과에 나타난다
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        List<Map<String, Object>> availableSites = response.jsonPath().getList("$");
-        
-        // A로 시작하는 사이트는 대형
-        availableSites.forEach(site -> {
-            String siteNumber = (String) site.get("siteNumber");
-            if (siteNumber != null && !siteNumber.isEmpty()) {
-                assertThat(siteNumber).startsWith("A");
-            }
-        });
-    }
-    
-    @DisplayName("시나리오: 이름과 전화번호로 내 예약 조회")
+    @DisplayName("이름과 전화번호로 예약을 조회하면, 해당 고객의 모든 예약 목록이 반환된다")
     @Test
     void scenario_QueryMyReservationsByNameAndPhone() {
         /*
@@ -406,34 +327,28 @@ class CampingReservationAcceptanceTest extends TestBase {
         // Given: 고객 "김철수"가 전화번호 "010-1234-5678"로 여러 예약을 완료했다
         String customerName = "김철수";
         String phoneNumber = "010-1234-5678";
-        
+
         // 첫 번째 예약
         createReservation(customerName, phoneNumber, "1", LocalDate.now().plusDays(5), LocalDate.now().plusDays(7));
-        
+
         // 두 번째 예약
         createReservation(customerName, phoneNumber, "2", LocalDate.now().plusDays(10), LocalDate.now().plusDays(12));
-        
+
         // When: 이름과 전화번호로 예약을 조회한다
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .queryParam("name", customerName)
-                .queryParam("phone", phoneNumber)
-                .when().get("/api/reservations/my")
-                .then().log().all()
-                .extract();
-        
+        ExtractableResponse<Response> response = getMyReservations(customerName, phoneNumber);
+
         // Then: 해당 고객의 모든 예약 목록이 반환된다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         List<Map<String, Object>> reservations = response.jsonPath().getList("$");
         assertThat(reservations).hasSizeGreaterThanOrEqualTo(2);
-        
+
         reservations.forEach(reservation -> {
             assertThat(reservation.get("customerName")).isEqualTo(customerName);
             assertThat(reservation.get("phoneNumber")).isEqualTo(phoneNumber);
         });
     }
-    
-    @DisplayName("시나리오: 월별 예약 현황 캘린더 조회")
+
+    @DisplayName("년도, 월, 사이트ID로 캘린더를 조회하면, 예약된 날짜와 가능한 날짜가 구분되어 표시된다")
     @Test
     void scenario_MonthlyReservationCalendar() {
         /*
@@ -449,25 +364,16 @@ class CampingReservationAcceptanceTest extends TestBase {
         // Given: 사이트에 해당 월 예약이 있다
         Long siteId = 1L;
         LocalDate reservationDate = LocalDate.now().plusDays(10);
-        
+
         createReservation("김철수", "010-1234-5678", siteId.toString(), reservationDate, reservationDate.plusDays(1));
-        
+
         // When: 해당 월 사이트의 캘린더를 조회한다
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .queryParam("year", reservationDate.getYear())
-                .queryParam("month", reservationDate.getMonthValue())
-                .queryParam("siteId", siteId)
-                .when().get("/api/reservations/calendar")
-                .then().log().all()
-                .extract();
-        
+        ExtractableResponse<Response> response = getReservationCalendar(reservationDate.getYear(), reservationDate.getMonthValue(), siteId);
+
         // Then: 예약된 날짜와 가능한 날짜가 구분되어 표시된다
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        assertThat(response.jsonPath().get("year")).isEqualTo(reservationDate.getYear());
-        assertThat(response.jsonPath().get("month")).isEqualTo(reservationDate.getMonthValue());
-        assertThat(response.jsonPath().get("siteId")).isEqualTo(siteId.intValue());
+        assertThat(response.jsonPath().getInt("year")).isEqualTo(reservationDate.getYear());
+        assertThat(response.jsonPath().getInt("month")).isEqualTo(reservationDate.getMonthValue());
+        assertThat(response.jsonPath().getInt("siteId")).isEqualTo(siteId.intValue());
     }
-    
-    // Helper methods are now inherited from TestBase
 }
